WhiskerRAG/server/requirements.txt
fastapi
uvicorn
pydantic
httpx
pytest
pytest-asyncio
python-dotenv
pydantic-settings